+++ bbc1/core/bbc_config.py
@@ -160,5 +160,6 @@ class BBcConfig:
                 'storage_type': StorageType.FILESYSTEM,
                 'storage_path': None,
                 'advertise_in_domain0': False,
+                'max_body_size': bbclib.DEFAULT_MAX_BODY_SIZE,
             }
         return dc['asset_group_ids'][asset_group_str]