+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change log
 ======
 
+## v0.8
+* Implement system statistics API #8
+* Notification of transaction insertion is implemented #9
+* NAT traversal support (only for simple port forwarding) #11
+* user_id based transaction search #16
+* Windows support for libbbcsig library
+
 ## v0.7.4
 * Fix issues regarding ledger_subsystem #26, #27
 * Refactor ledger_subsystem